+++ test_target_lock_v1024.py
@@ -23,7 +23,7 @@ class Root:
 
 class TargetLockV1024Tests(unittest.TestCase):
     def test_version(self):
-        self.assertEqual(APP_VERSION,'1.0.144-rc4')
+        self.assertEqual(APP_VERSION,'1.0.144-rc5')
         self.assertEqual(FILE_VERSION,'1.0.144')
 
     def make_app(self):